2022.09.22_web_app_trainer/requirements_cristian.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
bnlearn==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
community==1.0.0b1
cryptography==37.0.1
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
df2onehot==1.0.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.2
fonttools==4.33.3
fsspec==2022.3.0
funcsigs==1.0.2
h11==0.13.0
idna==3.3
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
ismember==1.0.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonpickle==2.1.0
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.8
notebook==6.4.11
numpy==1.22.3
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pgmpy==0.1.18
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pypickle==1.1.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyvis==0.2.0
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
selenium==4.1.3
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
statsmodels==0.13.2
tabulate==0.8.9
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
torch==1.11.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
wget==3.2
widgetsnbextension==3.6.0
wsproto==1.1.0
